Extract specific links SQL queries in traits
These specific methods were polluting the file and made things harder to
read and maintain. I'm not sure it's a good solution to relocate them
into 6 different files, but I feel better for now :).
